Add files via upload
This .py program can process the .dat files that the measurement type "aging" of the Natezalnica creates.
It extracts the oscillation times, absolute times, lower and upper forces and lower and upper lengths of all files in a specified folder (when using on a new computer the path to the folder should be manually adjusted). This program also draws plots of the extracted files with respect to the current number of repetition.
Program works correctly with the Spyder4 editor. It might need a 'plt.show()' at the end of each drawing function to work with some other editors.