First version
This program processes the .dat file made by the Natezalnica when choosing the 'aging' measurement type.
It draws the graphs of upper and lower forces, upper and lower lengths, oscillation times and calculates the Young's module, also drawing it as a graph. All graphs are drawn with respect to the number of repetitions. The program can merge measurements from different files if for some reason the measurement was done in more than one part. The directory to the files should be changed in line 158. The files should be named in such a manner that they are in the directory in the correct order (eg. 1,2,3...).