Version 1
This program processes the .dat files made by Natezalnica when choosing the 'relaxation' measurement type. It extracts the forces measured every 0.1 seconds and draws a graph showing how the force changes with time. The program can merge measurements from different files but they have to be in the specified directory (line 36) and the files should be in order in the directory (eg. 1,2,3..).

TODO:
1.) correct scale on x-axis (instead of the unit being '0.1 seconds'
2.) a marker when the translator started and stopped to move and in which direction
3.)