Still version 1
This program draws the measurements and the fitted curve it calculates on different graphs for either the whole range, only the upper relaxation or only the lower relaxation. It also calculates and prints out the parameters for the exponential rise/decay function.
To do: fix x axis units and fix Tau units to seconds.