some updates, still w.i.p.
Second update:
the characteristic times of exponential growth/fall are now calculated and printed in seconds. The average of all characteristic times for a set of files is also calculated and printed out.
Files can now have a header with every line starting with the '#' symbol and the program can still extract the correct data.