Version 2.0
This is a refined program accounting for the changes (more importantly, simplifications) that are now present in the measuring program as well as the .dat file it produces.
The functioning of the program is slightly different but it does the same thing as it used to. Uploaded as a pull request to a separate branch to keep the old version in case it turns it's actually useful.